Update to v220
Now that the app has been ported to Python3 in this new version,
we can move it away from the old runtimes that use python2.

Switch to the GNOME SDK as this includes the relevant versions
of PyGObject and pycairo.

Update sugar-toolkit-gtk3 to python3 version, which also requires
intltool from an updatd shared-modules pull.

Update python-dbus and decorator and and use python3 for both modules.

Remove upstreamed patches.

I build from the current master, which is the same as v220 but
additionally includes some fixes that were required to make the new
app version run and produce acceptable appdata.