Update to GNOME 47 runtime
Add the necessary requirements.txt to regenerate
python3-requirements.json to the repo, along with a new
build-requirements.txt to generate python3-build-dependencies.json.

Add instructions to a README on how to use these files to regenerate the
Python module manifests.

Add a new non-Python build-dependency on patchelf.

Adjust the numpy test cleanup rule to be independent of the runtime's
Python version.

Update shared-modules.